requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-chroma>=0.1.0
huggingface-hub>=0.20.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0